Handle python 3.12 upgrade smoother + make utils file
